refactor(ai): migrate AI content generation methods to ContentAiUtil for better modularity
